sandcastle: cmr #305 fix: preserve fractional pay arrears

+++ ming_sim/db.py
@@ -4426,7 +4426,7 @@ class GameDB:
             # arrears 累计欠饷万两，按月应发归一成"欠饷月数*10"（截至 100），与各 0-100 短板相加。
             # 用浮点归一（排序键，不截断小数；线上 gemini）；arrears 虽 NOT NULL 仍 `or 0` 防御。
             pay = self._army_pay(r)
-            arr = int(r["arrears"] or 0)
+            arr = float(r["arrears"] or 0)
             arrears_norm = min(100.0, arr * 10.0 / pay) if pay > 0 else 0.0
             danger = (arrears_norm + (100 - int(r["supply"])) + (100 - int(r["morale"]))
                       + (100 - int(r["loyalty"])) + (100 - int(r["training"])))
@@ -4453,7 +4453,7 @@ class GameDB:
                     "morale": int(row["morale"]),
                     "training": int(row["training"]),
                     "equipment": int(row["equipment"]),
-                    "arrears": int(row["arrears"]),
+                    "arrears": float(row["arrears"]),
                     "mobility": int(row["mobility"]),
                     "loyalty": int(row["loyalty"]),
                     "firearm_equipment": int(row["firearm_equipment"]),
